Move SizeT/VectorSizeT to use uint64_t instead of size_t
The motivation for this change and the follow-up commit that actually moves to use uint64_t instead of size_t is two-fold:

1. size_t has a platform-dependent bit width. Since we're compiling for other architectures, it makes more sense for us to use explicitly sized types.

2. This clears the way for adding other MemberTypes and ElemKinds which are typed and explicitly sized, which is consistent with ONNX and Caffe2.

+++ examples/char-rnn.cpp
@@ -88,7 +88,7 @@ static void loadText(Tensor &inputText, Tensor &nextChar, llvm::StringRef text,
   auto S = idim[1];
 
   auto IH = inputText.getHandle();
-  auto NH = nextChar.getHandle<size_t>();
+  auto NH = nextChar.getHandle<uint64_t>();
 
   // Fill the tensor with slices from the sentence with an offset of 1.
   // Example:
